use checkboxes for table column show/hide
